タグの削除の時、nullの場合は早期returnするように修正

+++ app/src/application/tagApplicationService.ts
@@ -61,6 +61,8 @@ export default class TagApplicationService {
   public async delete(id: string): Promise<void> {
     const tag = await this.tagRepository.findById(id);
 
+    if (tag === null) return;
+
     await this.tagRepository.delete(tag);
   }
